pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ag-network"
version = "0.1.0"
description = "Agent network: Workflow orchestration for agentic AI with a multipurpose skillset"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "AG Network Team"}]
license = {text = "MIT"}

dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]
llm = [
    "anthropic>=0.18.0",
    "openai>=1.12.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
openai = [
    "openai>=1.12.0",
]
all = [
    "ag-network[dev,llm]",
]

[project.scripts]
ag = "agnetwork.cli:app"

[tool.setuptools]
packages = ["agnetwork"]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "C901",   # mccabe complexity
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
